requirements.txt
flask
PyPDF2
pandas
python-dotenv
google-genai
pdf2image
pytesseract
openpyxl